// web/src/apis/models/Body_upload_and_pends_api_content_upload_and_pend_post.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Body_upload_and_pends_api_content_upload_and_pend_post = {
    /**
     * The UID of the content to be created.
     */
    uid: string;
    /**
     * The file content to be created.
     */
    file: Blob;
    /**
     * The language of the audio file.
     */
    audio_language?: (string | null);
};