// web/src/apis/models/CreateContentRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CreateContentRequest = {
    /**
     * The URL of the content to be created.
     */
    url?: (string | null);
    /**
     * The file content to be created.
     */
    file?: (Blob | null);
};